requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==3.0.1